// src/app/shared/models/factura-Admi.model.ts
// factura.model.ts
export interface FacturasAdmi {
    id: number;
    identification: string;  
    
    names: string;    
    email: string;      
    number: string;
    total: number;
    created_at: string;          
    prefix: string;          
    reference_code: string; 
    status: number;
       
         
  }
  